refactor: move acceptance tests to the equinix_test package
