+++ equinix/data_source_metal_vlan_acc_test.go
@@ -1,16 +1,15 @@
-package equinix
+package equinix_test
 
 import (
 	"fmt"
-	"reflect"
 	"testing"
 
+	"github.com/equinix/terraform-provider-equinix/internal/acceptance"
 	"github.com/equinix/terraform-provider-equinix/internal/config"
 
 	"github.com/hashicorp/terraform-plugin-testing/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/hashicorp/terraform-plugin-testing/terraform"
-	"github.com/packethost/packngo"
 )
 
 func TestAccDataSourceMetalVlan_byVxlanFacility(t *testing.T) {
@@ -18,9 +17,9 @@ func TestAccDataSourceMetalVlan_byVxlanFacility(t *testing.T) {
 	fac := "sv15"
 
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:          func() { testAccPreCheck(t) },
-		ExternalProviders: testExternalProviders,
-		Providers:         testAccProviders,
+		PreCheck:          func() { acceptance.TestAccPreCheck(t) },
+		ExternalProviders: acceptance.TestExternalProviders,
+		Providers:         acceptance.TestAccProviders,
 		CheckDestroy:      testAccMetalDatasourceVlanCheckDestroyed,
 		Steps: []resource.TestStep{
 			{
@@ -65,9 +64,9 @@ func TestAccDataSourceMetalVlan_byVxlanMetro(t *testing.T) {
 	metro := "sv"
 
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:          func() { testAccPreCheck(t) },
-		ExternalProviders: testExternalProviders,
-		Providers:         testAccProviders,
+		PreCheck:          func() { acceptance.TestAccPreCheck(t) },
+		ExternalProviders: acceptance.TestExternalProviders,
+		Providers:         acceptance.TestAccProviders,
 		CheckDestroy:      testAccMetalDatasourceVlanCheckDestroyed,
 		Steps: []resource.TestStep{
 			{
@@ -136,9 +135,9 @@ func TestAccDataSourceMetalVlan_byVlanId(t *testing.T) {
 	metro := "sv"
 
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:          func() { testAccPreCheck(t) },
-		ExternalProviders: testExternalProviders,
-		Providers:         testAccProviders,
+		PreCheck:          func() { acceptance.TestAccPreCheck(t) },
+		ExternalProviders: acceptance.TestExternalProviders,
+		Providers:         acceptance.TestAccProviders,
 		CheckDestroy:      testAccMetalDatasourceVlanCheckDestroyed,
 		Steps: []resource.TestStep{
 			{
@@ -182,9 +181,9 @@ func TestAccDataSourceMetalVlan_byProjectId(t *testing.T) {
 	metro := "sv"
 
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:          func() { testAccPreCheck(t) },
-		ExternalProviders: testExternalProviders,
-		Providers:         testAccProviders,
+		PreCheck:          func() { acceptance.TestAccPreCheck(t) },
+		ExternalProviders: acceptance.TestExternalProviders,
+		Providers:         acceptance.TestAccProviders,
 		CheckDestroy:      testAccMetalDatasourceVlanCheckDestroyed,
 		Steps: []resource.TestStep{
 			{
@@ -223,106 +222,8 @@ data "equinix_metal_vlan" "dsvlan" {
 `, projSuffix, metro, desc)
 }
 
-func TestMetalVlan_matchingVlan(t *testing.T) {
-	type args struct {
-		vlans     []packngo.VirtualNetwork
-		vxlan     int
-		projectID string
-		facility  string
-		metro     string
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    *packngo.VirtualNetwork
-		wantErr bool
-	}{
-		{
-			name: "MatchingVLAN",
-			args: args{
-				vlans:     []packngo.VirtualNetwork{{VXLAN: 123}},
-				vxlan:     123,
-				projectID: "",
-				facility:  "",
-				metro:     "",
-			},
-			want:    &packngo.VirtualNetwork{VXLAN: 123},
-			wantErr: false,
-		},
-		{
-			name: "MatchingFac",
-			args: args{
-				vlans:    []packngo.VirtualNetwork{{FacilityCode: "fac"}},
-				facility: "fac",
-			},
-			want:    &packngo.VirtualNetwork{FacilityCode: "fac"},
-			wantErr: false,
-		},
-		{
-			name: "MatchingMet",
-			args: args{
-				vlans: []packngo.VirtualNetwork{{MetroCode: "met"}},
-				metro: "met",
-			},
-			want:    &packngo.VirtualNetwork{MetroCode: "met"},
-			wantErr: false,
-		},
-		{
-			name: "SecondMatch",
-			args: args{
-				vlans: []packngo.VirtualNetwork{{FacilityCode: "fac"}, {MetroCode: "met"}},
-				metro: "met",
-			},
-			want:    &packngo.VirtualNetwork{MetroCode: "met"},
-			wantErr: false,
-		},
-		{
-			name: "TwoMatches",
-			args: args{
-				vlans: []packngo.VirtualNetwork{{MetroCode: "met"}, {MetroCode: "met"}},
-				metro: "met",
-			},
-			want:    nil,
-			wantErr: true,
-		},
-		{
-			name: "ComplexMatch",
-			args: args{
-				vlans: []packngo.VirtualNetwork{{VXLAN: 987, FacilityCode: "fac", MetroCode: "skip"}, {VXLAN: 123, FacilityCode: "fac", MetroCode: "met"}, {VXLAN: 456, FacilityCode: "fac", MetroCode: "nope"}},
-				metro: "met",
-			},
-			want:    &packngo.VirtualNetwork{VXLAN: 123, FacilityCode: "fac", MetroCode: "met"},
-			wantErr: false,
-		},
-		{
-			name: "NoMatch",
-			args: args{
-				vlans:     nil,
-				vxlan:     123,
-				projectID: "pid",
-				facility:  "fac",
-				metro:     "met",
-			},
-			want:    nil,
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := matchingVlan(tt.args.vlans, tt.args.vxlan, tt.args.projectID, tt.args.facility, tt.args.metro)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("matchingVlan() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("matchingVlan() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
 func testAccMetalDatasourceVlanCheckDestroyed(s *terraform.State) error {
-	client := testAccProvider.Meta().(*config.Config).Metal
+	client := acceptance.TestAccProvider.Meta().(*config.Config).Metal
 
 	for _, rs := range s.RootModule().Resources {
 		if rs.Type != "equinix_metal_vlan" {
