+++ equinix/data_source_network_platform_acc_test.go
@@ -1,9 +1,10 @@
-package equinix
+package equinix_test
 
 import (
 	"fmt"
 	"testing"
 
+	"github.com/equinix/terraform-provider-equinix/internal/acceptance"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 )
 
@@ -16,8 +17,8 @@ func TestAccDataSourceNetworkDevicePlatform_basic(t *testing.T) {
 	}
 	resourceName := fmt.Sprintf("data.equinix_network_device_platform.%s", context["resourceName"].(string))
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:  func() { testAccPreCheck(t) },
-		Providers: testAccProviders,
+		PreCheck:  func() { acceptance.TestAccPreCheck(t) },
+		Providers: acceptance.TestAccProviders,
 		Steps: []resource.TestStep{
 			{
 				Config: testAccDataSourceNetworkDevicePlatformConfig_basic(context),
