+++ equinix/helper_test_test.go
@@ -1,0 +1,43 @@
+package equinix_test
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestHelper_slicesMatch(t *testing.T) {
+	// given
+	input := [][][]string{
+		{
+			{"DC", "SV", "FR"},
+			{"FR", "SV", "DC"},
+		},
+		{
+			{"SV"},
+			{},
+		},
+		{
+			{"DC", "DC", "DC"},
+			{"DC", "SV", "DC"},
+		},
+		{
+			{}, {},
+		},
+	}
+	expected := []bool{
+		true,
+		false,
+		false,
+		true,
+	}
+	// when
+	results := make([]bool, len(expected))
+	for i := range input {
+		results[i] = slicesMatch(input[i][0], input[i][1])
+	}
+	// then
+	for i := range expected {
+		assert.Equal(t, expected[i], results[i])
+	}
+}
