+++ equinix/provider.go
@@ -3,7 +3,6 @@ package equinix
 import (
 	"context"
 	"fmt"
-	"strings"
 	"time"
 
 	"github.com/equinix/terraform-provider-equinix/internal/resources/metal/metal_connection"
@@ -245,52 +244,4 @@ func isEmpty(v interface{}) bool {
 	default:
 		return false
 	}
-}
-
-func slicesMatch(s1, s2 []string) bool {
-	if len(s1) != len(s2) {
-		return false
-	}
-	visited := make([]bool, len(s1))
-	for i := 0; i < len(s1); i++ {
-		found := false
-		for j := 0; j < len(s2); j++ {
-			if visited[j] {
-				continue
-			}
-			if s1[i] == s2[j] {
-				visited[j] = true
-				found = true
-				break
-			}
-		}
-		if !found {
-			return false
-		}
-	}
-	return true
-}
-
-func slicesMatchCaseInsensitive(s1, s2 []string) bool {
-	if len(s1) != len(s2) {
-		return false
-	}
-	visited := make([]bool, len(s1))
-	for i := 0; i < len(s1); i++ {
-		found := false
-		for j := 0; j < len(s2); j++ {
-			if visited[j] {
-				continue
-			}
-			if strings.EqualFold(s1[i], s2[j]) {
-				visited[j] = true
-				found = true
-				break
-			}
-		}
-		if !found {
-			return false
-		}
-	}
-	return true
 }
