+++ equinix/resource_metal_project_api_key_acc_test.go
@@ -1,9 +1,10 @@
-package equinix
+package equinix_test
 
 import (
 	"fmt"
 	"testing"
 
+	"github.com/equinix/terraform-provider-equinix/internal/acceptance"
 	"github.com/equinix/terraform-provider-equinix/internal/config"
 
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
@@ -12,9 +13,9 @@ import (
 
 func TestAccMetalProjectAPIKey_basic(t *testing.T) {
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:          func() { testAccPreCheck(t) },
-		ExternalProviders: testExternalProviders,
-		Providers:         testAccProviders,
+		PreCheck:          func() { acceptance.TestAccPreCheck(t) },
+		ExternalProviders: acceptance.TestExternalProviders,
+		Providers:         acceptance.TestAccProviders,
 		CheckDestroy:      testAccMetalProjectAPIKeyCheckDestroyed,
 		Steps: []resource.TestStep{
 			{
@@ -46,7 +47,7 @@ resource "equinix_metal_project_api_key" "test" {
 }
 
 func testAccMetalProjectAPIKeyCheckDestroyed(s *terraform.State) error {
-	client := testAccProvider.Meta().(*config.Config).Metal
+	client := acceptance.TestAccProvider.Meta().(*config.Config).Metal
 	for _, rs := range s.RootModule().Resources {
 		if rs.Type != "equinix_metal_project_api_key" {
 			continue
